+++ build.gradle
@@ -11,7 +11,7 @@ buildscript {
     ext.kotlin_version = '1.3.0'
     ext.android_components_version = '0.40.0'
     ext.lifecycle_version = '1.1.1'
-    ext.navigation_version = '1.0.0-rc01'
+    ext.navigation_version = '1.0.0-rc02'
     ext.rxbinding_version = '2.2.0'
     ext.coroutine_version = '1.0.1'
     ext.androidxTest_version = '1.1.0'
